requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.7.3
numpy==1.26.4
packaging==24.2
pandas==2.0.3
pillow==11.1.0
plotly==5.18.0
pycparser==2.22
pyotp==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
robin-stocks==3.3.1
seaborn==0.12.2
six==1.17.0
tenacity==9.0.0
tzdata==2025.1
urllib3==2.3.0
